+++ src/shared/posts/1/241010_int_char_bool_string.ts
@@ -1,4 +1,4 @@
-import { Post } from ".";
+import { Post } from "..";
 
 const post: Post = {
   title: `int, char, bool, string`,
