+++ src/shared/posts/1/241014_queue_priority_queue.ts
@@ -1,4 +1,4 @@
-import { Post } from ".";
+import { Post } from "..";
 
 const post: Post = {
   title: "Queue / Priority Queue",
