+++ src/shared/posts/1/241014_vector.ts
@@ -1,4 +1,4 @@
-import { Post } from ".";
+import { Post } from "..";
 
 const post: Post = {
   title: `Vector`,
