+++ src/shared/posts/index.ts
@@ -4,17 +4,36 @@ export interface Post {
     tag: string;
     content: string;
 }
-import post000000_contents from './000000_contents';
-import post241010_input_output from './241010_input_output';
-import post241010_int_char_bool_string from './241010_int_char_bool_string';
-import post241014_queue_priority_queue from './241014_queue_priority_queue';
-import post241014_vector from './241014_vector';
+import post1_241010_input_output from './1/241010_input_output';
+import post1_241010_int_char_bool_string from './1/241010_int_char_bool_string';
+import post1_241014_queue_priority_queue from './1/241014_queue_priority_queue';
+import post1_241014_vector from './1/241014_vector';
 
 export const posts = [
-    post000000_contents,
-post241010_input_output,
-post241010_int_char_bool_string,
-post241014_queue_priority_queue,
-post241014_vector
+    [post1_241010_input_output, post1_241010_int_char_bool_string, post1_241014_queue_priority_queue, post1_241014_vector],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[],
+[]
 ];
-
